requirements.txt
pandas
sqlalchemy
psycopg2-binary
openpyxl
python-dotenv
python-dateutil
schedule